Added user validator to the login module
